backend/accounts/urls.py
from django.urls import path, include
from rest_framework import routers

from accounts import views

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet, basename='users')

urlpatterns = [
    path('', include(router.urls)),
    path('auth/webhook/', views.auth_webhook_handler, name='auth-webhook'),
]
