backend/requirements.txt
boto3==1.40.11
celery==5.5.3
crewai==0.159.0
crewai-tools==0.62.0
Django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
docker==7.1.0
ipython==9.4.0
markdown-it-py==4.0.0
redis==6.4.0
drf-spectacular==0.28.0
psycopg2==2.9.10